==> Codes/Step2.DEGs.Rmd <==
---
title: "Step2.DEGs"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo                  = TRUE
                      ,warning              = FALSE
                      ,message              = FALSE
                      ,error                = TRUE
                      ,dev                  = c('png','pdf')
                      ,pdf.options(encoding = "ISOLatin9.enc", useKerning=FALSE, useDingbats=FALSE)
)                      
```

```{r library}
library(Seurat)
library(ggplot2)
library(ggpubr)
library(clusterProfiler)
library(org.Hs.eg.db)
library(AnnotationHub)
library(EnsDb.Hsapiens.v86)
library(ggplot2)
library(DESeq2)
```

```{r loading data}
load("./dir/Github/Step1.Data_processing.Rdata")
```

```{r Pseudobulk}
BC.Se.mal <- SplitObject(BC.Se.mal, split.by = "Patient")

# Creating the pseudobulk data by summing up the gene counts for each patient

HER2 = data.frame(matrix("", ncol = 0, nrow = 29733))  
for (nums in c("CID45171","CID4066","CID3921")) {
  HER2 <- cbind(as.data.frame(HER2),rowSums(BC.Se.mal[[nums]]@assays$RNA@counts[1:29733,]))
}

colnames(HER2) <- c("HER2_CID45171","HER2_CID4066","HER2_CID3921")

ER = data.frame(matrix("", ncol = 0, nrow = 29733))  
for (nums in c("CID4535","CID4530N","CID4471","CID4463","CID4461","CID4290A","CID4067","CID3948","CID3941")) {
  ER <- cbind(as.data.frame(ER),rowSums(BC.Se.mal[[nums]]@assays$RNA@counts[1:29733,]))
}

colnames(ER) <- c("ER_CID4535","ER_CID4530N","ER_CID4471", "ER_CID4463","ER_CID4461","ER_CID4290A","ER_CID4067","ER_CID3948","ER_CID3941")


TNBC = data.frame(matrix("", ncol = 0, nrow = 29733))  
for (nums in c("CID4523","CID4515","CID4513","CID44991","CID44971","CID4495","CID4465","CID3963")) {
  TNBC <- cbind(as.data.frame(TNBC),rowSums(BC.Se.mal[[nums]]@assays$RNA@counts[1:29733,]))
}

colnames(TNBC) <- c("TNBC_CID4523","TNBC_CID4515","TNBC_CID4513", "TNBC_CID44991", "TNBC_CID44971", "TNBC_CID4495", "TNBC_CID4465", "TNBC_CID3963")

Pseudo_Bulk <- cbind(HER2,ER,TNBC)
```

```{r DESeq2}
desc.de <- data.frame(id= colnames(Pseudo_Bulk),
                       conditions=factor((c(rep(("HER2"),3),rep(("ER"),9), rep(("TNBC"),8))),
                                   levels=c("HER2","ER","TNBC")))
                      
dds_cond <- DESeqDataSetFromMatrix(countData = Pseudo_Bulk, colData = desc.de, design = ~conditions)

dds_cond <- DESeq(dds_cond)


TNBCvsER <- as.data.frame(results(dds_cond, contrast = c("conditions","TNBC", "ER"), pAdjustMethod="fdr"))
TNBCvsHER2 <- as.data.frame(results(dds_cond, contrast = c("conditions","TNBC", "HER2"), pAdjustMethod="fdr"))
HER2vsER <- as.data.frame(results(dds_cond, contrast = c("conditions","HER2", "ER"), pAdjustMethod="fdr"))

TNBCvsER <- TNBCvsER[order(TNBCvsER$padj),]
TNBCvsHER2 <- TNBCvsHER2[order(TNBCvsHER2$padj),]
HER2vsER <- HER2vsER[order(HER2vsER$padj),]
```


````{r volcano plot,echo=FALSE, fig.show='hold', out.width="100%", fig.fullwidth = TRUE, fig.height=4, fig.width=4}
with(TNBCvsER, plot(log2FoldChange, -log10(padj), pch=20, main="DEGs TNBC vs ER",xlim=c(-15,15), ylim=c(0,22),cex.axis = 1.3,cex.lab = 1.3,cex.main=1.8,col="#666666",bty='n'))
with(subset(TNBCvsER , padj<.05), points(log2FoldChange, -log10(padj), pch=20,col="#666666"))
with(subset(TNBCvsER, padj<.05 & (log2FoldChange)>2), points(log2FoldChange, -log10(padj), pch=20, col="#FFFF33"))
with(subset(TNBCvsER, padj<.05 & (log2FoldChange)<=-2), points(log2FoldChange, -log10(padj), pch=20, col="#880088"))

legend("topleft",
       legend=c("Upregulated","Downregulated"),
       col=c("#FFCCFF","#FFFF99"),
       lty=1,lwd=2, box.lty = 0, box.lwd = 1,cex = 1)

with(TNBCvsHER2, plot(log2FoldChange, -log10(padj), pch=20, main="DEGs TNBC vs HER2",xlim=c(-15,15), ylim=c(0,22),cex.axis = 1.3,cex.lab = 1.3,cex.main=1.8,col="#666666",bty='n'))
with(subset(TNBCvsHER2 , padj<.05), points(log2FoldChange, -log10(padj), pch=20,col="#666666"))
with(subset(TNBCvsHER2, padj<.05 & (log2FoldChange)>2), points(log2FoldChange, -log10(padj), pch=20, col="#FFFF33"))
with(subset(TNBCvsHER2, padj<.05 & (log2FoldChange)<=-2), points(log2FoldChange, -log10(padj), pch=20, col="#880088"))

legend("topleft",
       legend=c("Upregulated","Downregulated"),
       col=c("#FFCCFF","#FFFF99"),
       lty=1,lwd=2, box.lty = 0, box.lwd = 1,cex = 1)


with(HER2vsER, plot(log2FoldChange, -log10(padj), pch=20, main="DEGs HER2 vs ER",xlim=c(-15,15), ylim=c(0,22),cex.axis = 1.3,cex.lab = 1.3,cex.main=1.8,col="#666666",bty='n'))
with(subset(HER2vsER , padj<.05), points(log2FoldChange, -log10(padj), pch=20,col="#666666"))
with(subset(HER2vsER, padj<.05 & (log2FoldChange)>2), points(log2FoldChange, -log10(padj), pch=20, col="#FFFF33"))
with(subset(HER2vsER, padj<.05 & (log2FoldChange)<=-2), points(log2FoldChange, -log10(padj), pch=20, col="#880088"))

legend("topleft",
       legend=c("Upregulated","Downregulated"),
       col=c("#FFCCFF","#FFFF99"),
       lty=1,lwd=1, box.lty = 1, box.lwd = 0,cex = 1)
```


```{r GO,fig.show='hold', out.width="50%", fig.fullwidth = TRUE,fig.width=8, fig.height=12}
TNBCvsER_up <- subset(TNBCvsER, subset = padj < 0.05 & log2FoldChange >2  )
TNBCvsER_down <- subset(TNBCvsER, subset = padj < 0.05 & log2FoldChange <= -2 )
TNBCvsHER2_up <- subset(TNBCvsHER2, subset = padj < 0.05 & log2FoldChange >2  )
TNBCvsHER2_down <- subset(TNBCvsHER2, subset = padj < 0.05 & log2FoldChange <= -2 )
HER2vsER_up <- subset(HER2vsER, subset = padj < 0.05 & log2FoldChange >2  )
HER2vsER_down <- subset(HER2vsER, subset = padj < 0.05 & log2FoldChange <= -2 )

# MalvsNorm_up <- subset(MalvsNorm, subset = padj < 0.05 & log2FoldChange > 0 )


TNBCvsER_up_GO <- enrichGO(gene = rownames(TNBCvsER_up),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

TNBCvsER_down_GO <- enrichGO(gene = rownames(TNBCvsER_down),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

TNBCvsHER2_up_GO <- enrichGO(gene = rownames(TNBCvsHER2_up),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

TNBCvsHER2_down_GO <- enrichGO(gene = rownames(TNBCvsHER2_down),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

HER2vsER_up_GO <- enrichGO(gene = rownames(HER2vsER_up),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

HER2vsER_down_GO <- enrichGO(gene = rownames(HER2vsER_down),
                   OrgDb = org.Hs.eg.db,
                   keyType = "SYMBOL",
                   ont = "CC",
                   pAdjustMethod = "BH",
                   pvalueCutoff = 0.05,
                   qvalueCutoff = 0.05)

TNBCvsER_up_GO@result$Log.adj <- -log10(TNBCvsER_up_GO@result$p.adjust)
barplot(TNBCvsER_up_GO,showCategory = 5, font.size = 16, x = "Log.adj") +theme_classic()+ggtitle("GO - TNBCvsER \nUpregulated genes") +  theme(title =element_text(size=16, face='bold'),
 legend.position="none", 
 axis.text=element_text(size=14,colour = "black")) + geom_bar(stat="identity", fill = "lightblue")+scale_x_continuous(breaks=seq(0, 15, 5))+ xlab("-log(adj.pvalue)")

TNBCvsER_down_GO@result$Log.adj <- -log10(TNBCvsER_down_GO@result$p.adjust)
barplot(TNBCvsER_down_GO,showCategory = 5, font.size = 16, x = "Log.adj")+theme_classic()+ggtitle("GO - TNBCvsER \nDownregulated genes") +  theme(title =element_text(size=16, face='bold'),
 legend.position="none", 
 axis.text=element_text(size=14,colour = "black")) + geom_bar(stat="identity", fill = "lightblue")+scale_x_continuous(breaks=seq(0, 15, 5))+ xlab("-log(adj.pvalue)")

TNBCvsHER2_up_GO@result$Log.adj <- -log10(TNBCvsHER2_up_GO@result$p.adjust)
barplot(TNBCvsHER2_up_GO,showCategory = 5, font.size = 16, x = "Log.adj")+theme_classic()+ggtitle("GO - TNBCvsHER2 \nUpregulated genes") +  theme(title =element_text(size=16, face='bold',colour = "black"),
 legend.position="none", 
 axis.text=element_text(size=14,colour = "black")) + geom_bar(stat="identity", fill = "lightblue")+scale_x_continuous(breaks=seq(0, 20, 5))+ xlab("-log(adj.pvalue)")

```

```{r, save}
save(Pseudo_Bulk, file = "./dir/Github/Step2.Pseudobulk.Rdata")
```
